refactor | select space or rent
